Resolving Issues - The CartItems where not mapped to the Order while creating the Order, and the CartItem's Ordered status wasn't updated

+++ order-service/src/main/java/com/example/order/model/Order.java
@@ -3,6 +3,7 @@ package com.example.order.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -23,9 +24,9 @@ public class Order {
     private List<CartItem> orderItems;
 
     private String paymentMethod;
-    private Double totalAmount;
-    private Double deliveryCost;
-    private Double totalPayableAmount;
+    private BigDecimal totalAmount;
+    private BigDecimal deliveryCost;
+    private BigDecimal totalPayableAmount;
 
     // Reference to the customer record
     private Long customerId;
